Module Homomorphisms

Module Homomorphisms

Abstract Algebra provides homomorphisms of finitely presented modules.

Generic functionality

The following generic functionality is provided for module homomorphisms.

Constructors

Homomorphisms of AbstractAlgebra modules, $f : R^s \to R^t$, can be represented by $s\times t$ matrices over $R$.

AbstractAlgebra.Generic.ModuleHomomorphism — Method.

ModuleHomomorphism(M1::AbstractAlgebra.FPModule{T}, M2::AbstractAlgebra.FPModule{T}, m::AbstractAlgebra.MatElem{T}) where T <: RingElement

Create the homomorphism $f : M_1 \to M_2$ represented by the matrix $m$.

Examples

M = FreeModule(ZZ, 2)
f = ModuleHomomorphism(M, M, matrix(ZZ, 2, 2, [1, 2, 3, 4]))

m = M([ZZ(1), ZZ(2)])

f(m)

Kernels

AbstractAlgebra.Generic.kernel — Method.

kernel(f::ModuleHomomorphism{T}) where T <: RingElement

Returns a pair K, f consisting of the kernel object $K$ of the given module homomorphism $f$ (as a submodule of its domain) and the canonical injection from the kernel into the domain of $f$

Examples

M = FreeModule(ZZ, 3)

m = M([ZZ(1), ZZ(2), ZZ(3)])

S, f = Submodule(M, [m])
Q, g = QuotientModule(M, S)

kernel(g)