Generic univariate polynomials over a noncommutative ring

Generic univariate polynomials over a noncommutative ring

AbstractAlgebra.jl provides a module, implemented in src/generic/NCPoly.jl for generic polynomials over any noncommutative ring belonging to the AbstractAlgebra abstract type hierarchy.

As well as implementing the Univariate Polynomial interface, there are many additional generic algorithms implemented for such polynomial rings. We describe this generic functionality below.

All of the generic functionality is part of a submodule of AbstractAlgebra called Generic. This is exported by default so that it is not necessary to qualify the function names with the submodule name.

Types and parent objects

Polynomials implemented using the AbstractAlgebra generics have type Generic.NCPoly{T} where T is the type of elements of the coefficient ring. Internally they consist of a Julia array of coefficients and some additional fields for length and a parent object, etc. See the file src/generic/GenericTypes.jl for details.

Parent objects of such polynomials have type Generic.NCPolyRing{T}.

The string representation of the variable of the polynomial ring and the base/coefficient ring $R$ is stored in the parent object.

The polynomial element types belong to the abstract type AbstractAlgebra.NCPolyElem{T} and the polynomial ring types belong to the abstract type AbstractAlgebra.NCPolyRing{T}. This enables one to write generic functions that can accept any AbstractAlgebra polynomial type.

Note that both the generic polynomial ring type Generic.NCPolyRing{T} and the abstract type it belongs to, AbstractAlgebra.NCPolyRing{T} are both called NCPolyRing. The former is a (parameterised) concrete type for a polynomial ring over a given base ring whose elements have type T. The latter is an abstract type representing all polynomial ring types in AbstractAlgebra.jl, whether generic or very specialised (e.g. supplied by a C library).

Polynomial ring constructors

In order to construct polynomials in AbstractAlgebra.jl, one must first construct the polynomial ring itself. This is accomplished with the following constructor.

PolynomialRing(R::AbstractAlgebra.NCRing, s::AbstractString; cached::Bool = true)

Given a base ring R and string s specifying how the generator (variable) should be printed, return a tuple S, x representing the new polynomial ring $S = R[x]$ and the generator $x$ of the ring. By default the parent object S will depend only on R and x and will be cached. Setting the optional argument cached to false will prevent the parent object S from being cached.

A shorthand version of this function is provided: given a base ring R, we abbreviate the constructor as follows.

R["x"]

Here are some examples of creating polynomial rings and making use of the resulting parent objects to coerce various elements into the polynomial ring.

Examples

R = MatrixAlgebra(ZZ, 2)
S, x = PolynomialRing(R, "x")
T, y = PolynomialRing(T, "y")
U, z = R["z"]

f = S()
g = S(123)
h = T(BigInt(1234))
k = T(x + 1)
m = U(z + 1)

All of the examples here are generic polynomial rings, but specialised implementations of polynomial rings provided by external modules will also usually provide a PolynomialRing constructor to allow creation of their polynomial rings.

Basic ring functionality

Once a polynomial ring is constructed, there are various ways to construct polynomials in that ring.

The easiest way is simply using the generator returned by the PolynomialRing constructor and build up the polynomial using basic arithmetic, as described in the Ring interface.

The Julia language also has special syntax for the construction of polynomials in terms of a generator, e.g. we can write 2x instead of 2*x.

The polynomial rings in AbstractAlgebra.jl implement the full Ring interface. Of course the entire Univariate Polynomial Ring interface is also implemented.

We give some examples of such functionality.

Examples

R = MatrixAlgebra(ZZ, 2)
S, x = PolynomialRing(R, "x")
T, y = PolynomialRing(T, "y")

f = x^3 + 3x + 21
g = (x + 1)*y^2 + 2x + 1

h = zero(T)
k = one(S)
isone(k) == true
iszero(f) == false
n = length(g)
U = base_ring(T)
V = base_ring(y + 1)
v = var(T)
U = parent(y + 1)
g == deepcopy(g)
t = divexact(2g, 2)

Polynomial functionality provided by AbstractAlgebra.jl

The functionality listed below is automatically provided by AbstractAlgebra.jl for any polynomial module that implements the full Univariate Polynomial Ring interface over a noncommutative ring. This includes AbstractAlgebra.jl's own generic polynomial rings.

But if a C library provides all the functionality documented in the Univariate Polynomial Ring interface over a noncommutative ring, then all the functions described here will also be automatically supplied by AbstractAlgebra.jl for that polynomial type.

Of course, modules are free to provide specific implementations of the functions described here, that override the generic implementation.

Basic functionality

AbstractAlgebra.Generic.lead — Method.

lead(a::Generic.PolynomialElem)

Return the leading coefficient of the given polynomial. This will be the nonzero coefficient of the term with highest degree unless the polynomial in the zero polynomial, in which case a zero coefficient is returned.

AbstractAlgebra.Generic.trail — Method.

trail(a::Generic.PolynomialElem)

Return the trailing coefficient of the given polynomial. This will be the nonzero coefficient of the term with lowest degree unless the polynomial in the zero polynomial, in which case a zero coefficient is returned.

AbstractAlgebra.Generic.gen — Method.

gen(R::AbstractAlgebra.NCPolyRing)

Return the generator of the given polynomial ring.

AbstractAlgebra.Generic.isgen — Method.

isgen(a::Generic.PolynomialElem)

Return true if the given polynomial is the constant generator of its polynomial ring, otherwise return false.

AbstractAlgebra.Generic.isunit — Method.

isunit(a::Generic.PolynomialElem)

Return true if the given polynomial is a unit in its polynomial ring, otherwise return false.

AbstractAlgebra.Generic.ismonomial — Method.

ismonomial(a::Generic.PolynomialElem)

Return true if the given polynomial is a monomial.

AbstractAlgebra.Generic.isterm — Method.

isterm(a::Generic.PolynomialElem)

Return true if the given polynomial has one term. This function is recursive, with all scalar types returning true.

Examples

R = MatrixAlgebra(ZZ, 2)
S, x = PolynomialRing(R, "x")
T, y = PolynomialRing(S, "y")

a = zero(T)
b = one(T)

c = BigInt(1)*z^2 + BigInt(1)
d = x*y^2 + (x + 1)*y + 3

f = lead(d)
y = gen(T)
g = isgen(y)
m = isunit(b)
n = degree(d)
isterm(2y^2) == true
ismonomial(y^2) == true

Truncation

Base.truncate — Method.

truncate(a::Generic.PolynomialElem, n::Int)

Return $a$ truncated to $n$ terms.

AbstractAlgebra.Generic.mullow — Method.

mullow(a::AbstractAlgebra.NCPolyElem{T}, b::AbstractAlgebra.NCPolyElem{T}, n::Int) where {T <: NCRingElem}

Return $a\times b$ truncated to $n$ terms.

Examples

R = MatrixAlgebra(ZZ, 2)
S, x = PolynomialRing(R, "x")
T, y = PolynomialRing(S, "y")

f = x*y^2 + (x + 1)*y + 3
g = (x + 1)*y + (x^3 + 2x + 2)

h = truncate(f, 1)
k = mullow(f, g, 4)

Reversal

Base.reverse — Method.

reverse(x::Generic.PolynomialElem, len::Int)

Return the reverse of the polynomial $x$, thought of as a polynomial of the given length (the polynomial will be notionally truncated or padded with zeroes before the leading term if necessary to match the specified length). The resulting polynomial is normalised. If len is negative we throw a DomainError().

Base.reverse — Method.

reverse(x::Generic.PolynomialElem)

Return the reverse of the polynomial $x$, i.e. the leading coefficient of $x$ becomes the constant coefficient of the result, etc. The resulting polynomial is normalised.

Examples

R = MatrixAlgebra(ZZ, 2)
S, x = PolynomialRing(R, "x")
T, y = PolynomialRing(S, "y")

f = x*y^2 + (x + 1)*y + 3

g = reverse(f, 7)
h = reverse(f)

Shifting

AbstractAlgebra.Generic.shift_left — Method.

shift_left(f::Generic.PolynomialElem, n::Int)

Return the polynomial $f$ shifted left by $n$ terms, i.e. multiplied by $x^n$.

AbstractAlgebra.Generic.shift_right — Method.

shift_right(f::Generic.PolynomialElem, n::Int)

Return the polynomial $f$ shifted right by $n$ terms, i.e. divided by $x^n$.

Examples

R = MatrixAlgebra(ZZ, 2)
S, x = PolynomialRing(R, "x")
T, y = PolynomialRing(S, "y")

f = x*y^2 + (x + 1)*y + 3

g = shift_left(f, 7)
h = shift_right(f, 2)

Evaluation

AbstractAlgebra.Generic.evaluate — Method.

evaluate(a::AbstractAlgebra.NCPolyElem, b::T) where {T <: NCRingElem}

Evaluate the polynomial $a$ at the value $b$ and return the result.

AbstractAlgebra.Generic.evaluate — Method.

evaluate(a::AbstractAlgebra.NCPolyElem, b::Union{Integer, Rational, AbstractFloat})

Evaluate the polynomial $a$ at the value $b$ and return the result.

We also overload the functional notation so that the polynomial $f$ can be evaluated at $a$ by writing $f(a)$.

Examples

R = MatrixAlgebra(ZZ, 2)
S, x = PolynomialRing(R, "x")
T, y = PolynomialRing(S, "y")
   
f = x*y^2 + (x + 1)*y + 3

k = evaluate(f, 3)
m = evaluate(f, x^2 + 2x + 1)
r = f(23)

Derivative

AbstractAlgebra.Generic.derivative — Method.

derivative(a::Generic.PolynomialElem)

Return the derivative of the polynomial $a$.

Examples

R = MatrixAlgebra(ZZ, 2)
S, x = PolynomialRing(R, "x")
T, y = PolynomialRing(S, "y")

f = x*y^2 + (x + 1)*y + 3

h = derivative(f)