Submodules

Submodules

AbstractAlgebra allows the construction of submodules/subvector spaces of AbstractAlgebra modules over euclidean domains. These are given as the submodule generated by a finite list of elements in the original module.

We define two submodules to be equal if they are (transitively) submodules of the same module $M$ and their generators generate the same set of elements.

As well as implementing the entire Module interface, AbstractAlgebra submodules also provide the following interface.

Constructors

AbstractAlgebra.Submodule — Method.

Submodule(m::Module{T}, gens::Vector{<:ModuleElem{T}}) where T <: RingElement

Return the submodule S of the module m generated by the given generators, given as elements of m, and a map which is the canonical injection from S to m.

AbstractAlgebra.Subspace — Method.

Subspace(m::Module{T}, gens::Vector{<:ModuleElem{T}}) where T <: FieldElement

Return the subspace of the vector space m generated by the given generators, given as elements of m, and a map which is the canonical injection from S to m.

Examples

M = FreeModule(ZZ, 2)

m = M([ZZ(1), ZZ(2)])
n = M([ZZ(2), ZZ(-1)])

N, f = Submodule(M, [m, n])

v = N([ZZ(3), ZZ(4)])
v2 = f(v)

V = VectorSpace(QQ, 2)

m = V([QQ(1), QQ(2)])
n = V([QQ(2), QQ(-1)])

N, f = Subspace(V, [m, n])

Functionality for submodules

In addition to the Module interface, AbstractAlgebra submodules implement the following functionality.

Basic manipulation

AbstractAlgebra.Generic.supermodule — Method.

supermodule(M::Submodule{T}) where T <: RingElement

Return the module that this module is a submodule of.

AbstractAlgebra.Generic.issubmodule — Method.

issubmodule(M::AbstractAlgebra.FPModule{T}, N::AbstractAlgebra.FPModule{T}) where T <: RingElement

Return true if $N$ was constructed as a submodule of $M$. The relation is taken transitively (i.e. subsubmodules are submodules for the purposes of this relation, etc). The module $M$ is also considered a submodule of itself for this relation.

AbstractAlgebra.Generic.iscompatible — Method.

iscompatible(M::AbstractAlgebra.FPModule{T}, N::AbstractAlgebra.FPModule{T}) where T <: RingElement

Return true, P if the given modules are compatible, i.e. that they are (transitively) submodules of the same module, P. Otherwise return false, M.

Examples

M = FreeModule(ZZ, 2)
m = M([ZZ(2), ZZ(3)])
n = M([ZZ(1), ZZ(4)])
N1 = Submodule(M, [m, n])
N2 = Submodule(M, [m])

supermodule(N1) == M
iscompatible(N1, N2)
issubmodule(M, N)

Intersection

Base.intersect — Method.

Base.intersect(M::AbstractAlgebra.FPModule{T}, N::AbstractAlgebra.FPModule{T}) where T <: RingElement

Return the intersection of the modules $M$ as a submodule of $M$. Note that $M$ and $N$ must be (constructed as) submodules (transitively) of some common module $P$.

Examples

M = FreeModule(ZZ, 2)
m = M([ZZ(2), ZZ(3)])
n = M([ZZ(1), ZZ(4)])
N1 = Submodule(M, [m, n])
N2 = Submodule(M, [m])

I = intersect(N1, N2)